--- lib/actions/shared.types.ts
@@ -1,11 +1,0 @@
-export type UpdateUserParams = {
-  userId: string;
-  updateData: {
-    name?: string;
-    username?: string;
-    portfolioWebsite?: string;
-    location?: string;
-    bio?: string;
-  };
-  path: string;
-};
